+++ src/components/pokemon/PokemonItemHeader.vue
@@ -6,15 +6,33 @@
   >
     <img
       class="location"
-      src="@/assets/pokemons/location.jpg"
+      :src="
+        require(`@/assets/locations/${pokemonHabitatsBackground.get(habitat)}`)
+      "
       :height="locationHeight"
       :width="locationWidth"
     />
-    <img v-if="!image" :src="silouette" alt="pokemon silouette" />
+    <div
+      v-if="!image"
+      class="translucent-circle"
+      :style="{
+        height: `${locationHeight / 2}px`,
+        width: `${locationHeight / 2}px`,
+        top: `${locationHeight / 2 - locationHeight / 4}px`,
+      }"
+    ></div>
     <img
-      v-else
+      v-if="image"
       class="pokemon-image"
       :src="image"
+      alt="pokemon"
+      ref="image"
+      @load="setLocationHeight"
+    />
+    <img
+      v-else
+      class="small-pokemon-image"
+      :src="smallImage"
       alt="pokemon"
       ref="image"
       @load="setLocationHeight"
@@ -26,8 +44,8 @@
 
 <script>
 import debounce from 'lodash/debounce';
-import silouette from '@/assets/pokemons/silouette.png';
 import { capitalizeWord } from '@/lib/helpers';
+import { pokemonHabitatsBackground } from '@/constants/pokemonHabitatsBackground';
 
 export default {
   name: 'PokemonItemHeader',
@@ -35,13 +53,16 @@ export default {
     return {
       locationHeight: 0,
       locationWidth: 0,
-      silouette,
+      location: '',
+      pokemonHabitatsBackground,
     };
   },
   props: {
     image: {
       type: String,
-      required: true,
+    },
+    smallImage: {
+      type: String,
     },
     name: {
       type: String,
@@ -49,6 +70,10 @@ export default {
     },
     topPosition: {
       type: Number,
+      required: true,
+    },
+    habitat: {
+      type: String,
       required: true,
     },
   },
@@ -91,6 +116,8 @@ export default {
 
     @media (min-width: $min-width-fourth-break) {
       border-radius: 2rem;
+      border: 0.2rem solid var(--secondary-border-color);
+      box-shadow: var(--main-box-shadow);
       max-width: 24rem;
     }
 
@@ -120,6 +147,13 @@ export default {
     }
   }
 
+  .translucent-circle {
+    position: absolute;
+    background-color: rgba(255, 255, 255, 0.6);
+    border-radius: 50%;
+    z-index: 3;
+  }
+
   .pokemon-image {
     width: 20rem;
     z-index: 5;
@@ -134,14 +168,29 @@ export default {
     }
 
     @media (min-width: $min-width-fourth-break) {
-      border: 0.2rem solid var(--secondary-border-color);
-      border-radius: 2rem;
-      box-shadow: var(--main-box-shadow);
       max-width: 20rem;
     }
 
     @media (min-width: $min-width-fifth-break) {
       max-width: 30rem;
+    }
+  }
+
+  .small-pokemon-image {
+    width: 10rem;
+    z-index: 5;
+    padding: 2rem;
+
+    @media (min-width: $min-width-first-break) {
+      width: 15rem;
+    }
+
+    @media (min-width: $min-width-fourth-break) {
+      max-width: 10rem;
+    }
+
+    @media (min-width: $min-width-fifth-break) {
+      max-width: 15rem;
     }
   }