+++ src/lib/localStorage.js
@@ -1,4 +1,5 @@
 const DARKMODE_ENABLED = 'darkmodeEnabled';
+const GUESSES_IN_A_ROW = 'guessesInARow';
 
 export function isDarkModeEnabled() {
   return localStorage.getItem(DARKMODE_ENABLED) === 'true';
@@ -7,4 +8,12 @@ export function isDarkModeEnabled() {
 export function toggleDarkMode() {
   const isDarkModeEnabled = localStorage.getItem(DARKMODE_ENABLED) === 'true';
   localStorage.setItem(DARKMODE_ENABLED, !isDarkModeEnabled);
+}
+
+export function getGuessesInARow() {
+  return Number(localStorage.getItem(GUESSES_IN_A_ROW));
+}
+
+export function setGuessesInARow(guessesInARow) {
+  localStorage.setItem(GUESSES_IN_A_ROW, guessesInARow);
 }