+++ tests/unit/views/PageNotFoundView.test.js
@@ -1,6 +1,14 @@
 import { shallowMount } from '@vue/test-utils';
 import PageNotFoundView from '@/views/PageNotFoundView.vue';
 
+jest.mock('@/lib/store', () => ({
+  state: {
+    store: {
+      isDarkmodeEnabled: false,
+    },
+  },
+}));
+
 describe('PageNotFoundView', () => {
   it('renders the BasePage component with correct title and content', () => {
     const wrapper = shallowMount(PageNotFoundView);
@@ -19,5 +27,8 @@ describe('PageNotFoundView', () => {
     const span = basePage.find('.page-not-found-text');
     expect(span.exists()).toBe(true);
     expect(span.text()).toBe("How did you get here? There's nothing to see!");
+
+    const img = basePage.find('img');
+    expect(img.exists()).toBe(true);
   });
 });